+++ CHANGELOG.md
@@ -34,6 +34,8 @@
 - (Feature) Pre OOM Abort function
 - (Bugfix) Fix ErrorArray String function
 - (Feature) Switch services to Port names
+- (Feature) Configurable ArangoD Port
+- (Feature) Allow to exclude metrics
 - (Feature) Do not restart member if all pods in group are not ready
 
 ## [1.2.20](https://github.com/arangodb/kube-arangodb/tree/1.2.20) (2022-10-25)